benchmarks/ServerlessBench/case4-handler/requirements.txt
minio
Pillow
